Reorganizing sample app packages
