+++ sample-app/app/src/main/java/co/elastic/otel/android/sample/network/CityWeatherService.kt
@@ -1,6 +1,6 @@
-package co.elastic.apm.android.sample.network
+package co.elastic.otel.android.sample.network
 
-import co.elastic.apm.android.sample.network.data.ForecastResponse
+import co.elastic.otel.android.sample.network.data.ForecastResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
