+++ sample-app/app/src/main/java/co/elastic/otel/android/sample/network/data/CurrentWeatherResponse.kt
@@ -1,3 +1,3 @@
-package co.elastic.apm.android.sample.network.data
+package co.elastic.otel.android.sample.network.data
 
 data class CurrentWeatherResponse(val temperature: Double)
