+++ sample-app/app/src/main/java/co/elastic/otel/android/sample/network/data/ForecastResponse.kt
@@ -1,4 +1,4 @@
-package co.elastic.apm.android.sample.network.data
+package co.elastic.otel.android.sample.network.data
 
 import com.google.gson.annotations.SerializedName
